Authentication and Authorization Using JWT Successful.

+++ controllers/api/v1/posts_api.js
@@ -22,29 +22,30 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    post.deleteOne();
+    if (post.user == req.user.id) {
+      post.deleteOne();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    // if (req.xhr) {
-    //   return res.status(200).json({
-    //     data: {
-    //       post_id: req.params.id,
-    //     },
-    //     message: "Post Deleted Successfully",
-    //   });
-    // }
+      if (req.xhr) {
+        return res.status(200).json({
+          data: {
+            post_id: req.params.id,
+          },
+          message: "Post Deleted Successfully",
+        });
+      }
 
-    // req.flash("success", "Post and Associated Comments Deleted!");
+      // req.flash("success", "Post and Associated Comments Deleted!");
 
-    return res.status(200).json({
-      message: "Post and Associated Comments Deleted!",
-    });
-    // } else {
-    //   req.flash("error", "You Cannot Delete This Post!");
-    //   return res.redirect("back");
-    // }
+      return res.status(200).json({
+        message: "Post and Associated Comments Deleted!",
+      });
+    } else {
+      return res.status(401).json({
+        message: "You Cannot Delete This Post!",
+      });
+    }
   } catch (err) {
     console.log("***********", err);
     return res.status(500).json({
